+++ src/index.js
@@ -170,10 +170,14 @@ class Squiss extends EventEmitter {
   /**
    * Queues the given message for deletion. The message will actually be deleted from SQS per the settings
    * supplied to the constructor.
-   * @param {Message} msg The message to be deleted.
+   * @param {Message|string} msg The message object to be deleted, or the receipt handle of a message to be deleted
    */
   deleteMessage(msg) {
-    this._delQueue.push({ Id: msg.raw.MessageId, ReceiptHandle: msg.raw.ReceiptHandle })
+    if (msg instanceof Message) {
+      this._delQueue.push({ Id: msg.raw.MessageId, ReceiptHandle: msg.raw.ReceiptHandle })
+    } else {
+      this._delQueue.push({ Id: msg, ReceiptHandle: msg })
+    }
     this.handledMessage()
     if (this._delQueue.length >= this._opts.deleteBatchSize) {
       if (this._delTimer) {